third_party/blink/web_tests/http/tests/security/subresourceIntegrity/subresource-integrity-script-no-cors-no-xorigin-with-creds.html

<!DOCTYPE html>
<head>
    <title>SRI on cross origin CORS disabled but with credentials script, without CORS fetch</title>
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        setup({single_test: true});

        var result = '';
        var scriptLoaded = function() {
            assert_unreached('Script ran.');
            done();
        };
        var scriptError = function() {
            assert_equals(result, '', 'Script did not run.');
            done();
        };
    </script>
    </script>
    <script integrity="sha256-mk_DwEwGGoyrh4sgi9t0nYTO1Lrft3w6vzaNuMaZrtQ=" src="http://localhost:8000/security/resources/cors-script.php?credentials=true&cors=false&value=ran" onload="scriptLoaded();" onerror="scriptError()"></script>
</body>
</html>
