third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-script-no-cors-no-xorigin-with-creds-expected.txt

CONSOLE ERROR: Subresource Integrity: The resource 'http://localhost:8000/security/resources/cors-script.php?credentials=true&cors=false&value=ran' has an integrity attribute, but the resource requires CORS to be enabled to check the integrity, and it is not. The resource has been blocked.
This is a testharness.js-based test.
PASS SRI on cross origin CORS disabled but with credentials script, without CORS fetch 
Harness: the test ran to completion.
